chore(talkback): delete the dead, divergent AccessibilityFocusTracker
AccessibilityFocusTracker shipped a full class (getCurrentFocus,
buildTraversalOrder, findElementIndex, invalidate*, singleton, 5s cache) that was
never instantiated or called anywhere in src/. The live focus-navigation path
uses FocusElementMatcher + FocusPathCalculator instead; the only runtime exercise
was the tracker's own test (#3919).

It was not merely dead but divergent, so wiring it in would have silently
regressed matching: its matchesSelector did exact, case-sensitive equality
against camelCase fields (resourceId, contentDesc, testTag), while CtrlProxy
elements are parsed with hyphenated keys (resource-id, content-desc, test-tag).
FocusElementMatcher reads both key styles and matches partially and
case-insensitively. Its caches were also only cleared by the unused invalidate*
methods.

Delete the class and its test. The one live export, ElementSelector (plus its
SelectorBounds), moves to src/features/talkback/ElementSelector.ts — beside its
three consumers (FocusElementMatcher, FocusPathCalculator,
FocusNavigationExecutor) rather than stranded in utils.

Closes #3919
